pyproject.toml
[tool.poetry]
name = "whats_eat"
version = "0.1.0"
description = ""
authors = ["Oriol Burgaya <oriolbv@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6.8"
scipy = "^1.3"
matplotlib = "^3.1"
pandas = "^0.25.0"
statsmodels = "^0.10.1"
sklearn = "^0.0.0"
tensorflow = "^1.14"
keras = "^2.2"
image = "^1.5"
scikit-image = "^0.15.0"
Flask = "^1.1"

[tool.poetry.dev-dependencies]
